+++ src/main/java/org/galatea/starter/entrypoint/ApiError.java
@@ -1,4 +1,4 @@
-package org.galatea.starter.service;
+package org.galatea.starter.entrypoint;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
