+++ src/main/java/org/galatea/starter/entrypoint/RestExceptionHandler.java
@@ -1,8 +1,8 @@
-package org.galatea.starter.service;
+package org.galatea.starter.entrypoint;
 
 import lombok.extern.slf4j.Slf4j;
 
-import org.galatea.starter.domain.exception.EntityNotFoundException;
+import org.galatea.starter.entrypoint.exception.EntityNotFoundException;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
